Remove orphaned application and tooling code
Delete disconnected UI surfaces, standalone tools, and unused dependencies while preserving active comments and image optimization.
