+++ apps/desktop/src/chat/components/persistent-chat.test.tsx
@@ -22,7 +22,6 @@ const mocks = vi.hoisted(() => ({
     messages: [],
     onAddContextEntity: vi.fn(),
     onDraftContextRefsChange: vi.fn(),
-    onRemoveContextEntity: vi.fn(),
     pendingRefs: [],
     regenerate: vi.fn(),
     sendMessage: vi.fn(),
