+++ apps/web/package.json
@@ -22,69 +22,41 @@
     "@anlg/ui": "workspace:*",
     "@anlg/user-error": "workspace:*",
     "@anlg/utils": "workspace:*",
-    "@chenglou/pretext": "^0.0.4",
-    "@deepgram/sdk": "^4.11.3",
     "@floating-ui/dom": "^1.7.6",
     "@honeycombio/opentelemetry-web": "^1.3.0",
     "@iconify-icon/react": "^3.0.3",
-    "@mdx-js/mdx": "^3.1.1",
-    "@mdx-js/react": "^3.1.1",
-    "@mux/mux-player-react": "^3.11.7",
     "@nangohq/frontend": "^0.69.48",
     "@netlify/cache": "^3.4.4",
     "@netlify/vite-plugin-tanstack-start": "^1.3.3",
-    "@opentelemetry/api": "^1.9.1",
     "@opentelemetry/auto-instrumentations-web": "^0.49.1",
-    "@opentelemetry/resources": "^2.6.1",
     "@sentry/tanstackstart-react": "^10.48.0",
-    "@stripe/stripe-js": "^8.11.0",
     "@supabase/ssr": "^0.7.0",
     "@supabase/supabase-js": "^2.103.0",
     "@t3-oss/env-core": "^0.13.11",
     "@tailwindcss/vite": "^4.2.2",
-    "@tanstack/react-devtools": "^0.7.11",
     "@tanstack/react-form": "^1.29.0",
     "@tanstack/react-query": "^5.99.0",
-    "@tanstack/react-query-devtools": "^5.99.0",
     "@tanstack/react-router": "^1.168.19",
-    "@tanstack/react-router-devtools": "^1.166.13",
     "@tanstack/react-router-ssr-query": "^1.166.11",
     "@tanstack/react-start": "^1.167.33",
-    "@tanstack/router-plugin": "^1.167.19",
     "@unpic/react": "^1.0.2",
-    "@uppy/core": "^5.2.0",
-    "@uppy/tus": "^5.1.1",
-    "chart.js": "^4.5.1",
-    "clsx": "^2.1.1",
-    "dayjs": "^1.11.20",
-    "drizzle-orm": "^0.44.7",
-    "exa-js": "^1.10.2",
-    "github-slugger": "^2.0.0",
-    "html-to-image": "^1.11.13",
-    "jszip": "^3.10.1",
     "jwt-decode": "^4.0.0",
     "mdx-mermaid": "^2.0.3",
     "mermaid": "^11.16.1",
     "motion": "^11.18.2",
-    "postgres": "^3.4.9",
     "posthog-js": "^1.367.0",
     "react": "^19.2.5",
-    "react-chartjs-2": "^5.3.1",
     "react-dom": "^19.2.5",
-    "react-markdown": "^10.1.0",
     "react-tweet": "^3.3.0",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-slug": "^6.0.0",
     "remark-gfm": "^4.0.1",
-    "semver": "7.6.3",
     "sharp": "^0.35.3",
-    "shiki": "^3.23.0",
     "streamdown": "^2.5.0",
     "stripe": "^20.4.1",
     "tailwind-scrollbar-hide": "^4.0.0",
     "tailwindcss": "^4.2.2",
     "unpic": "^4.2.2",
-    "usehooks-ts": "^3.1.1",
     "vaul": "^1.1.2",
     "zod": "^4.5.4"
   },
@@ -95,13 +67,9 @@
     "@dotenvx/dotenvx": "^1.61.0",
     "@sentry/cli": "^3.6.1",
     "@tailwindcss/typography": "^0.5.19",
-    "@testing-library/dom": "^10.4.1",
-    "@testing-library/react": "^16.3.2",
-    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.19.17",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
-    "@types/semver": "^7.7.1",
     "@vitejs/plugin-react": "^6.0.1",
     "js-yaml": "^4.3.1",
     "jsdom": "^27.4.0",
@@ -110,7 +78,6 @@
     "tanstack-router-sitemap": "^1.0.13",
     "typescript": "^5.9.3",
     "vite": "^8.0.16",
-    "wrangler": "4.129.0",
-    "web-vitals": "^5.2.0"
+    "wrangler": "4.129.0"
   }
 }
