+++ apps/web/src/lib/shared-note-route-state.test.ts
@@ -3,10 +3,8 @@ import test from "node:test";
 
 import {
   getInvitationRouteFailure,
-  getLinkSharedNoteFallbackSnapshot,
   getLinkSharedNoteRouteGate,
 } from "./shared-note-route-state.ts";
-import type { SharedNoteSnapshot } from "./shared-notes.ts";
 
 test("keeps failed invitation acceptance retryable", () => {
   assert.equal(
@@ -67,35 +65,3 @@ test("authenticated access outranks continuation failures", () => {
     "continuation-error",
   );
 });
-
-test("link access provides an authenticated fallback without a token", () => {
-  const authenticatedSnapshot = sharedNoteSnapshot(7);
-  const linkSnapshot = sharedNoteSnapshot(6);
-
-  assert.equal(
-    getLinkSharedNoteFallbackSnapshot({
-      authenticatedSnapshot,
-      linkSnapshot: null,
-    }),
-    authenticatedSnapshot,
-  );
-  assert.equal(
-    getLinkSharedNoteFallbackSnapshot({
-      authenticatedSnapshot,
-      linkSnapshot,
-    }),
-    linkSnapshot,
-  );
-});
-
-function sharedNoteSnapshot(contentRevision: number): SharedNoteSnapshot {
-  return {
-    shareId: "00000000-0000-4000-8000-000000000001",
-    schemaVersion: 1,
-    contentRevision,
-    title: "Weekly sync",
-    body: { type: "doc" },
-    attachments: [],
-    publishedAt: "2026-07-17T12:00:00Z",
-  };
-}
