+++ apps/web/src/routes/share/$shareId.tsx
@@ -7,8 +7,8 @@ import {
 } from "@/components/shared-note-actions";
 import { SharedNoteChatPanel } from "@/components/shared-note-chat-panel";
 import { SharedNoteCollaboration } from "@/components/shared-note-collaboration";
+import { SharedNoteCollaborationViewer } from "@/components/shared-note-collaboration-viewer";
 import type { SharedAttachmentResolver } from "@/components/shared-note-document";
-import { SharedNoteEditableViewer } from "@/components/shared-note-editable-viewer";
 import {
   SharedNoteLoading,
   SharedNoteTransientError,
@@ -158,15 +158,12 @@ function Component() {
 
   return (
     <>
-      <SharedNoteEditableViewer
+      <SharedNoteCollaborationViewer
         key={snapshot.shareId}
         snapshot={snapshot}
         authenticatedNote={authenticatedNote}
-        fallbackAccessLabel={stableAccessLabel}
-        fallbackSnapshot={stableNote?.snapshot ?? snapshot}
         meetingMetadata={preview}
         resolveAttachment={resolveAttachment}
-        revokedBehavior="read-only"
         signedIn={user !== null}
         accessLabel={accessLabel}
         collaboration={
