--- packages/agent-designer/README.md
@@ -1,7 +1,0 @@
-```bash
-infisical run \
-  --projectId=87dad7b5-72a6-4791-9228-b3b86b169db1 \
-  --path="/slack-internal" \
-  --env=prod \
-  -- pnpm -F agent-designer dev
-```
