--- packages/agent-designer/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@anlg/agent-designer",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "dev": "npx @langchain/langgraph-cli dev",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@anlg/agent-core": "workspace:^",
-    "@langchain/core": "^1.1.39",
-    "@langchain/langgraph": "^1.2.8",
-    "@langchain/langgraph-checkpoint-postgres": "^1.0.1",
-    "@langchain/openai": "^1.4.4",
-    "@t3-oss/env-core": "^0.13.11",
-    "langsmith": "^0.4.12",
-    "pg": "^8.20.0",
-    "promptl-ai": "^0.4.12",
-    "zod": "^4.5.4"
-  },
-  "devDependencies": {
-    "@langchain/langgraph-cli": "^1.1.17",
-    "@types/bun": "^1.3.12",
-    "@types/pg": "^8.20.0",
-    "typescript": "^5.9.3"
-  }
-}
