--- packages/agent-designer/src/agent.ts
@@ -1,7 +1,0 @@
-import { env } from "@anlg/agent-core";
-
-import { graph } from "./graph";
-
-process.env.LANGSMITH_TRACING = env.LANGSMITH_API_KEY ? "true" : "false";
-
-export const agent = graph;
