--- packages/agent-designer/src/nodes/agent.ts
@@ -1,9 +1,0 @@
-import path from "path";
-
-import { createAgentNode, loadPrompt } from "@anlg/agent-core";
-
-import { tools } from "../tools";
-
-const prompt = loadPrompt(path.join(import.meta.dirname, ".."));
-
-export const agentNode = createAgentNode(prompt, tools);
