--- packages/agent-designer/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "include": ["src/**/*.ts"],
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": false,
-    "declaration": true,
-    "outDir": "dist",
-    "rootDir": "src",
-    "strict": true,
-    "skipLibCheck": true,
-    "noUncheckedSideEffectImports": true,
-    "types": ["@types/bun"]
-  }
-}
